backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic>=2.0
openai>=1.40.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
redis>=5.2.0
python-multipart>=0.0.9
boto3>=1.34.0
pymupdf>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
